Moved WFC panel to new format
